Change copyLarge file

+++ src/main/java/top/viewv/model/symmetric/CallBack.java
@@ -1,5 +1,5 @@
 package top.viewv.model.symmetric;
 
 public interface CallBack {
-    void report(int process);
+    void report(long process);
 }
