+++ src/main/java/top/viewv/model/symmetric/EncryptProgress.java
@@ -15,7 +15,7 @@ public class EncryptProgress implements CallBack {
     }
 
     @Override
-    public void report(int process){
+    public void report(long process){
         System.out.println("Callback Encryption Process: "+process);
     }
 }
